+++ modules/atom/include/distance.h
@@ -13,42 +13,81 @@
 #include <IMP/core/XYZ.h>
 #include "Hierarchy.h"
 #include "Selection.h"
+#include <IMP/algebra/distance.h>
 #include "IMP/base_types.h"
 
 IMPATOM_BEGIN_NAMESPACE
-
-//! Calculate the root mean square deviation between two sets of 3D points.
 /**
-   \note the function assumes correspondence between the two sets of
-   points and does not perform rigid alignment.
-
-   \genericgeometry
+   Get the rmsd between two lists of particles
  */
+inline double get_rmsd_transforming_first(const IMP::algebra::Transformation3D& tr,
+                                          const core::XYZs& s0,
+                                          const core::XYZs& s1
+                       ) {
+  return algebra::get_rmsd_transforming_first(tr, s0, s1);
+}
+
+/**
+   Get the rmsd between two lists of particles
+ */
+inline double get_rmsd(const core::XYZs& s0, const core::XYZs& s1) {
+  return algebra::get_rmsd(s0, s1);
+}
+
+/**
+   Get the rmsd between two lists of particles.
+ */
+IMPATOM_DEPRECATED_FUNCTION_DECL(2.2)
 template <class Vector3DsOrXYZs0, class Vector3DsOrXYZs1>
 inline double get_rmsd(const Vector3DsOrXYZs0& m1, const Vector3DsOrXYZs1& m2,
                        const IMP::algebra::Transformation3D& tr_for_second =
                            IMP::algebra::get_identity_transformation_3d()) {
-  IMP_USAGE_CHECK(std::distance(m1.begin(), m1.end()) ==
-                      std::distance(m2.begin(), m2.end()),
-                  "The input sets of XYZ points "
-                      << "should be of the same size");
-  double rmsd = 0.0;
-  typename Vector3DsOrXYZs0::const_iterator it0 = m1.begin();
-  typename Vector3DsOrXYZs1::const_iterator it1 = m2.begin();
-  for (; it0 != m1.end(); ++it0, ++it1) {
-    algebra::Vector3D tred =
-        tr_for_second.get_transformed(get_vector_d_geometry(*it1));
-    rmsd += algebra::get_squared_distance(get_vector_d_geometry(*it0), tred);
-  }
-  return std::sqrt(rmsd / m1.size());
+  IMPATOM_DEPRECATED_FUNCTION_DEF(2.2, "Use it in IMP::algebra");
+  return algebra::get_rmsd_transforming_first(tr_for_second, m2, m1);
+}
+
+
+/** RMSD on a pair of Selections.*/
+inline double get_rmsd(const Selection& s0, const Selection& s1) {
+  return algebra::get_rmsd(s0.get_selected_particles(), s1.get_selected_particles());
 }
 
 /** RMSD on a pair of Selections.*/
+inline double get_rmsd_transforming_first(const IMP::algebra::Transformation3D& tr,
+                                          const Selection& s0, const Selection& s1) {
+  return algebra::get_rmsd_transforming_first(tr, s0.get_selected_particles(),
+                                              s1.get_selected_particles());
+}
+/**
+   \deprecated_at{2.2} Use get_rmsd_transforming_first()
+ */
+IMPATOM_DEPRECATED_FUNCTION_DECL(2.2)
 inline double get_rmsd(const Selection& s0, const Selection& s1,
-                       const IMP::algebra::Transformation3D& tr_for_second =
-                           IMP::algebra::get_identity_transformation_3d()) {
-  return get_rmsd(s0.get_selected_particles(), s1.get_selected_particles(),
-                  tr_for_second);
+                       const algebra::Transformation3D& tr_for_second) {
+  IMPATOM_DEPRECATED_FUNCTION_DEF(2.2, "Use IMP::atom::get_rmsd_transforming_first()");
+  return get_rmsd_transforming_first(tr_for_second, s1, s0);
+}
+
+/**
+   \deprecated_at{2.2} Use IMP::algebra::get_rmsd_transforming_first() or
+   IMP::algebra::get_rmsd()
+ */
+IMPATOM_DEPRECATED_FUNCTION_DECL(2.2)
+inline double get_rmsd(const algebra::Vector3Ds& s0, const algebra::Vector3Ds& s1,
+                       const IMP::algebra::Transformation3D& tr_for_second) {
+  IMPATOM_DEPRECATED_FUNCTION_DEF(2.2, "Use IMP::algebra::get_rmsds()");
+  return algebra::get_rmsd_transforming_first(tr_for_second, s1, s0);
+}
+
+/**
+   \deprecated_at{2.2} Use IMP::algebra::get_rmsd_transforming_first() or
+   IMP::algebra::get_rmsd()
+ */
+IMPATOM_DEPRECATED_FUNCTION_DECL(2.2)
+inline double get_rmsd(const core::XYZs& s0, const core::XYZs& s1,
+                       const IMP::algebra::Transformation3D& tr_for_second) {
+  IMPATOM_DEPRECATED_FUNCTION_DEF(2.2, "Use IMP::algebra::get_rmsds()");
+  return algebra::get_rmsd_transforming_first(tr_for_second, s1, s0);
 }
 
 //! Calculate the root mean square deviation between two sets of 3D points.